Added interactions for Unity Viz mode to cue files for getting recommendations. RIght click in viz mode cues files for sending to the recommender system. Pressing R key sends the recommendation.
Added fixes to the version1.py script as well. Args is the entire first argument from OSC, not an array of arguments as originally thought. So, using args directly. Also currently sending an array of strings instead of a single comma separated string for recommendations. Integration tested.

This should work as soon as the mapping between file names and ES audio keys is obtained.